refactored get weight details to tech records service
